cyberscraper/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
fake-useragent==1.3.0
stem==1.8.1
flask==3.0.0
aiohttp==3.9.3
sentence-transformers==2.2.2
huggingface-hub==0.19.4  # Add specific version
numpy==1.26.4
xgboost==2.0.3
shap==0.45.1
pandas==2.2.1
scikit-learn==1.4.0
web3==6.15.1
eth-utils==2.3.1
eth-typing==3.5.2
ipfshttpclient==0.8.0a2
celery==5.3.6
google-generativeai==0.3.2
python-dotenv==1.0.0
textblob==0.17.1
spacy==3.7.2
selenium==4.16.0
gql==3.5.0
transformers==4.36.2
torch==2.1.2
lime==0.2.0.1
scipy>=1.9.0  # Required by LIME
joblib>=1.1.0
tqdm>=4.65.0